Changed scoping to file scope, changed project naming
