+++ ReviewsMicroservice.Application/Interfaces/IRepositories/IReviewsRepository.cs
