+++ ReviewsMicroservice.Application/ReviewsMicroservice.Application.csproj
@@ -4,10 +4,11 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>Application</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Shared\Shared.csproj" />
+      <ProjectReference Include="..\ReviewsMicroservice.Shared\ReviewsMicroservice.Shared.csproj" />
     </ItemGroup>
 
     <ItemGroup>
