+++ ReviewsMicroservice.Infrastructure/ReviewsMicroservice.Infrastructure.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>Infrastructure</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
@@ -15,8 +16,8 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Application\Application.csproj" />
-      <ProjectReference Include="..\Shared\Shared.csproj" />
+      <ProjectReference Include="..\ReviewsMicroservice.Application\ReviewsMicroservice.Application.csproj" />
+      <ProjectReference Include="..\ReviewsMicroservice.Shared\ReviewsMicroservice.Shared.csproj" />
     </ItemGroup>
 
 </Project>
