+++ ReviewsMicroservice.Shared/ReviewsMicroservice.Shared.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>Shared</RootNamespace>
     </PropertyGroup>
 
 </Project>
